List genes involved in present GPS for a specific pathway in the summary_results

This function lists the genes involved in the present GPS for a pathway of interest, odered by their contribution to the significance of the pathway.

getGenes(yy, i, idmap = load_data("idmap"))

Arguments

yy
A sigora analysis result object (created by sigora).

i
The rank position of the pathway of interest in summary_results.

idmap
A dataframe for converting between different gene-identifier types (e.g. ENSEMBL, ENTREZ and HGNC-Symbols of genes). Most users do not need to set this argument, as there is a built-in conversion table.

Value

A table (dataframe) with ids of the relevant genes, ranked by their contribution to the statistical significance of the pathway.

See also

Examples

data('kegH')
set.seed(seed=12345)
a1<-genesFromRandomPathways(kegH,3,50)
#> ### randomly selected pathways are: 
#> hsa04218 
#>  hsa00590 
#>  hsa04917 
## originally selected pathways:\cr
a1[["selectedPathways"]]
#> [1] "hsa04218" "hsa00590" "hsa04917"
## what are the genes
a1[["genes"]]
#>  [1] "9134"      "91860"     "5499"      "22800"     "1111"      "90550"    
#>  [7] "6776"      "1874"      "8792"      "4051"      "28996"     "22808"    
#> [13] "1557"      "890"       "5594"      "493869"    "5606"      "8600"     
#> [19] "2876"      "100137049" "83447"     "894"       "1030"      "1559"     
#> [25] "11145"     "4773"      "1573"      "4893"      "7248"      "5296"     
#> [31] "293"       "3708"      "2053"      "147746"    "874"       "1081"     
#> [37] "23291"     "3803"      "8681"      "292"       "2878"      "2678"     
#> [43] "391013"    "824"       "5605"      "3972"      "4616"      "85417"    
#> [49] "891"       "5291"     
## sigora's results with this input:\cr
sigoraRes <- sigora(GPSrepo =kegH, queryList = a1[["genes"]],level = 2)
#>   pathwy.id                 description    pvalues Bonferroni successes
#> 1  hsa04218         Cellular senescence 1.669e-108 5.090e-106     56.68
#> 2  hsa00590 Arachidonic acid metabolism  3.125e-41  9.531e-39     21.83
#>   PathwaySize        N sample.size
#> 1     2037.98 452219.2       87.58
#> 2      606.97 452219.2       87.58
## Genes related to the second most significant result:
getGenes(sigoraRes,2)
#>         gene contribution Ensembl.Gene.ID        Symbol
#> 1       4051        4.025 ENSG00000186529        CYP4F3
#> 2       2053        2.525 ENSG00000120915         EPHX2
#> 3        874        2.365 ENSG00000159231          CBR3
#> 4       2678        1.565 ENSG00000100031          GGT1
#> 5       2878        1.565 ENSG00000211445          GPX3
#> 6     493869        1.565 ENSG00000164294          GPX8
#> 7       2876        1.220 ENSG00000233276          GPX1
#> 8       1557        1.155 ENSG00000165841       CYP2C19
#> 9       1573        1.155 ENSG00000134716        CYP2J2
#> 10     11145        1.050 ENSG00000176485       PLA2G16
#> 11    391013        1.000 ENSG00000187980       PLA2G2C
#> 12      1559        0.890 ENSG00000138109        CYP2C9
#> 13 100137049        0.875 ENSG00000168970 JMJD7-PLA2G4B
#> 14      8681        0.875 ENSG00000168970 JMJD7-PLA2G4B