Highlight the relevant genes for a specific pathway in its pathway diagram

This function highlights the genes involved in the present GPS for a pathway of interest in its diagram. Please note that this functionality is only implemented for results of Reactome or KEGG based analyses.

getURL(yy, i)

Arguments

yy
A sigora analysis result object (created by sigora).

i
The rank position of the pathway of interest in summary_results.

Value

The URL of the pathway diagram, where the relevant genes from your original query list are highlighted.

See also

Examples

data('kegH')
set.seed(seed=12345)
a1<-genesFromRandomPathways(kegH,3,50)
#> ### randomly selected pathways are: 
#> hsa04218 
#>  hsa00590 
#>  hsa04917 
## originally selected pathways:\cr
a1[["selectedPathways"]]
#> [1] "hsa04218" "hsa00590" "hsa04917"
## what are the genes
a1[["genes"]]
#>  [1] "9134"      "91860"     "5499"      "22800"     "1111"      "90550"    
#>  [7] "6776"      "1874"      "8792"      "4051"      "28996"     "22808"    
#> [13] "1557"      "890"       "5594"      "493869"    "5606"      "8600"     
#> [19] "2876"      "100137049" "83447"     "894"       "1030"      "1559"     
#> [25] "11145"     "4773"      "1573"      "4893"      "7248"      "5296"     
#> [31] "293"       "3708"      "2053"      "147746"    "874"       "1081"     
#> [37] "23291"     "3803"      "8681"      "292"       "2878"      "2678"     
#> [43] "391013"    "824"       "5605"      "3972"      "4616"      "85417"    
#> [49] "891"       "5291"     
## sigora's results with this input:\cr
sigoraRes <- sigora(GPSrepo =kegH, queryList = a1[["genes"]],level = 2)
#>   pathwy.id                 description    pvalues Bonferroni successes
#> 1  hsa04218         Cellular senescence 1.669e-108 5.090e-106     56.68
#> 2  hsa00590 Arachidonic acid metabolism  3.125e-41  9.531e-39     21.83
#>   PathwaySize        N sample.size
#> 1     2037.98 452219.2       87.58
#> 2      606.97 452219.2       87.58
## Diagram for the most significant result, where the genes from our list are highlighted in red:
getURL(sigoraRes,1)
#> assuming this is a result from KEGG ... 
#> [1] "https://www.genome.jp/kegg-bin/show_pathway?hsa04218+HIPK2+HIPK4+MCU+KIR2DL2+E2F4+CCNB3+FBXW11+CAPN2+RRAS2+MRAS+TSC1+CHEK1+CCNB1+MAP2K3+CDKN2B+SLC25A5+SLC25A6+SLC25A31+CCNA2+NFATC2+PPP1CA+GADD45B+ITPR1+CALML4+CCNE2+CCND2+MAP2K2+NRAS+PIK3CB+PIK3R2+MAPK1"