NCI human gene-pathway associations.

PID-NCI human pathway repository, as a data frame with three columns corresponding to : pathwayId , pathwayName, gene. This is an example of the expected format for the input data to makeGPS.

Source

<https://github.com/NCIP/pathway-interaction-database/tree/master/download>

Details

This dataset is provided to illustrate how to create your own GPS repositories.nciTable is a dataframe with threecolumns corresponding to pathwayId, pathwayName and gene. Each row describes the association between an individual gene and a PID-NCI pathway. As you see in the examples, section, one can convert this dataframe to a GPS repository using makeGPS, and save the results for future reuse. Using the thus created GPS repository one can preform Signature Overrepresentation Analysis on lists of genes of interest.

Examples

data(nciTable)
nciH<-makeGPS(pathwayTable=load_data('nciTable'))
#> Time difference of 0.8909671 secs
data(idmap)
ils<-grep("^IL",idmap[,"Symbol"],value=TRUE)
ilnci<-sigora(queryList=ils,GPSrepo=nciH,level=3)
#> [1] "Mapped identifiers from Symbol  to  Ensembl.Gene.ID ..."
#>       pathwy.id                    description   pvalues Bonferroni successes
#> 1   il23pathway IL23-mediated signaling events 5.494e-64  1.049e-61     36.27
#> 2   il27pathway IL27-mediated signaling events 3.164e-34  6.043e-32     18.14
#> 3 il12_2pathway IL12-mediated signaling events 3.188e-12  6.089e-10     13.20
#> 4    il1pathway  IL1-mediated signaling events 1.115e-09  2.130e-07      8.42
#> 5  il4_2pathway  IL4-mediated signaling events 1.070e-05  2.044e-03      9.03
#>   PathwaySize        N sample.size
#> 1      172.95 46257.95       93.08
#> 2       65.51 46257.95       93.08
#> 3      420.16 46257.95       93.08
#> 4      156.05 46257.95       93.08
#> 5      687.89 46257.95       93.08