Traditional Overrepresentation Analysis.

Traditional Overrepresentation Analysis by hypergeometric test: pathways are treated as collections of individual genes and all genes are treated as equally informative. This function is provided for comparison of the results of traditional methods to Sigora.

ora(geneList, GPSrepo, idmap = load_data("idmap"))

Arguments

geneList
A vector containing the list of genes of interest (e.g. differentially expressed genes). Following Identifier types are supported: Gene Symbols, ENTREZ-IDs or ENSEMBL-IDs.

GPSrepo
A GPS-repository (either one of the provided precomputed GPS repositories) or one created by makeGPS.

idmap
A dataframe for converting between different gene-identifier types (e.g. ENSEMBL, ENTREZ and HGNC-Symbols of genes). Most users do not need to set this argument, as there is a built-in conversion table.

Value

A dataframe with individual gene ORA results.

Details

The primary purpose of makeGPS is to create a GPS repository. It does, however, retain the original "single gene"-"pathway" associations for the purpose of followup analyses, such as comparison of sigora-results to traditional methods. ora is an implementation of the traditional (individual gene) Overrepresentation Analysis.

See also

Examples

data(kegM)
## select 50 genes from 3 mouse pathways
set.seed(seed=345)
a1<-genesFromRandomPathways(kegM,3,50)
#> ### randomly selected pathways are: 
#> mmu03010 
#>  mmu04722 
#>  mmu05223 
## originally selected pathways:
a1[["selectedPathways"]]
#> [1] "mmu03010" "mmu04722" "mmu05223"
## compare to traditional methods results:
oraRes <- ora(a1[["genes"]],kegM)
dim(oraRes)
#> [1] 4 6
oraRes
#>   pathwyid                       description success pathwaySize   pvalues
#> 1 mmu03010                          Ribosome      30         175 8.131e-37
#> 2 mmu04722    Neurotrophin signaling pathway      18         121 3.392e-20
#> 3 mmu03008 Ribosome biogenesis in eukaryotes       7         116 1.296e-05
#> 4 mmu05223        Non-small cell lung cancer       5          66 8.420e-05
#>         Bonfer
#> 1 1.032637e-34
#> 2 4.307840e-18
#> 3 1.645920e-03
#> 4 1.069340e-02